Moved Aqueous to Solvent.py.
